Cleanup
This doesn't really change anything, just cleans up the code a bit.
