+++ ouroboros-consensus/demo-playground/Mock/TxSubmission.hs
@@ -18,15 +18,16 @@ import           Data.Void
 import           Options.Applicative
 import           System.IO (IOMode (..))
 
+import           Ouroboros.Network.MonadClass hiding (threadDelay)
+import           Ouroboros.Network.Node (NodeId (..))
+import           Ouroboros.Network.Serialise
+
 import           Ouroboros.Consensus.Crypto.Hash (ShortHash)
 import qualified Ouroboros.Consensus.Crypto.Hash as H
 import           Ouroboros.Consensus.Ledger.Abstract
 import qualified Ouroboros.Consensus.Ledger.Mock as Mock
 import           Ouroboros.Consensus.Node (NodeKernel (getExtLedgerState))
-import           Ouroboros.Consensus.Util (Condense (..))
-import           Ouroboros.Network.MonadClass hiding (threadDelay)
-import           Ouroboros.Network.Node (NodeId (..))
-import           Ouroboros.Network.Serialise
+import           Ouroboros.Consensus.Util.Condense
 
 import           Mock.Mempool (Mempool (..), consistent, mempoolInsert)
 import           Mock.Protocol
