+++ ouroboros-consensus/src/Ouroboros/Consensus/Protocol/ModChainSel.hs
@@ -26,11 +26,12 @@ module Ouroboros.Consensus.Protocol.ModChainSel (
 
 import           Data.Proxy (Proxy (..))
 
-import           Ouroboros.Consensus.Protocol.Abstract
-import           Ouroboros.Consensus.Util
 import           Ouroboros.Network.Block (HasHeader, Slot)
 import           Ouroboros.Network.Chain (Chain)
 import           Ouroboros.Network.Serialise
+
+import           Ouroboros.Consensus.Protocol.Abstract
+import           Ouroboros.Consensus.Util.Condense
 
 class OuroborosTag (Protocol c) => ChainSelection c where
