+++ ouroboros-consensus/src/Ouroboros/Consensus/Util.hs
@@ -1,45 +1,32 @@
-{-# LANGUAGE BangPatterns           #-}
-{-# LANGUAGE DeriveGeneric          #-}
-{-# LANGUAGE FlexibleContexts       #-}
-{-# LANGUAGE FlexibleInstances      #-}
-{-# LANGUAGE FunctionalDependencies #-}
-{-# LANGUAGE MultiParamTypeClasses  #-}
-{-# LANGUAGE RankNTypes             #-}
-{-# LANGUAGE ScopedTypeVariables    #-}
-{-# LANGUAGE TypeApplications       #-}
-{-# LANGUAGE TypeFamilies           #-}
-{-# LANGUAGE UndecidableInstances   #-}
+{-# LANGUAGE BangPatterns        #-}
+{-# LANGUAGE ConstraintKinds     #-}
+{-# LANGUAGE GADTs               #-}
+{-# LANGUAGE PolyKinds           #-}
+{-# LANGUAGE ScopedTypeVariables #-}
 
+-- | Miscellaneous utilities
 module Ouroboros.Consensus.Util (
-    -- * Miscellaneous
-    foldlM'
+    Dict(..)
+  , Some(..)
+  , foldlM'
   , repeatedly
   , repeatedlyM
   , chunks
   , byteStringChunks
   , lazyByteStringChunks
   , whenJust
-    -- * Pretty-printing
-  , Condense(..)
   ) where
 
 import qualified Data.ByteString as Strict
 import qualified Data.ByteString.Lazy as Lazy
-import           Data.List (foldl', intercalate)
-import           Data.Map (Map)
-import qualified Data.Map.Strict as Map
-import           Data.Proxy
-import           Data.Set (Set)
-import qualified Data.Set as Set
-import           Numeric.Natural
-import           Text.Printf (printf)
+import           Data.Kind (Constraint)
+import           Data.List (foldl')
 
-import           Ouroboros.Consensus.Util.HList (All, HList (..))
-import qualified Ouroboros.Consensus.Util.HList as HList
+data Dict (a :: Constraint) where
+    Dict :: a => Dict a
 
-{-------------------------------------------------------------------------------
-  Miscellaneous
--------------------------------------------------------------------------------}
+data Some (f :: k -> *) where
+    Exists :: f a -> Some f
 
 foldlM' :: forall m a b. Monad m => (b -> a -> m b) -> b -> [a] -> m b
 foldlM' f = go
@@ -71,44 +58,3 @@ lazyByteStringChunks n bs
 whenJust :: Applicative f => Maybe a -> (a -> f ()) -> f ()
 whenJust (Just x) f = f x
 whenJust Nothing _  = pure ()
-{-# INLINE whenJust #-}
-
-{-------------------------------------------------------------------------------
-  Condensed but human-readable output
--------------------------------------------------------------------------------}
-
-class Condense a where
-  condense :: a -> String
-
-instance Condense String where
-  condense = id
-
-instance Condense Int where
-  condense = show
-
-instance Condense Word where
-  condense = show
-
-instance Condense Natural where
-  condense = show
-
-instance Condense Rational where
-  condense = printf "%.8f" . (fromRational :: Rational -> Double)
-
-instance {-# OVERLAPPING #-} Condense [String] where
-  condense ss = "[" ++ intercalate "," ss ++ "]"
-
-instance {-# OVERLAPPABLE #-} Condense a => Condense [a] where
-  condense as = "[" ++ intercalate "," (map condense as) ++ "]"
-
-instance All Condense as => Condense (HList as) where
-  condense as = "(" ++ intercalate "," (HList.collapse (Proxy @Condense) condense as) ++ ")"
-
-instance (Condense a, Condense b) => Condense (a, b) where
-  condense (a, b) = condense (a :* b :* Nil)
-
-instance Condense a => Condense (Set a) where
-  condense = condense . Set.toList
-
-instance (Condense k, Condense a) => Condense (Map k a) where
-  condense = condense . Map.toList
